+++ src/app/result/[id]/meal-container.tsx
@@ -2,6 +2,7 @@ import { MealImage } from '@/src/app/result/[id]/meal-image';
 import { DetailedMealWithTranslations } from '@/src/types/api.types';
 import { MealIngredientContainer } from '@/src/app/result/[id]/meal-ingredient-container';
 import { MealGeneral } from '@/src/app/result/[id]/meal-general';
+import { MealCommentContainer } from '@/src/app/result/[id]/meal-comment-container';
 
 interface MealContainerProps {
     complexMealObject: DetailedMealWithTranslations;
@@ -15,6 +16,7 @@ export function MealContainer({ complexMealObject }: MealContainerProps) {
             <MealImage imgUrl={meal.imgUrl} title={meal.title} />
             <MealGeneral meal={meal} />
             <MealIngredientContainer ingredients={meal.ingredients} />
+            <MealCommentContainer />
         </>
     );
 }