+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.38.0] - 2024-08-27
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Capability to add the comments for the meals.
+
 ## [1.37.0] - 2024-08-27
 ### Changed
 - [Marek Kurańda](https://github.com/mjkuranda): Rendering Polish translations of ingredients at search and create page.